requirements.txt
discord.py[voice]
dotenv
flask
pynacl

# dev
pytest
pytest-asyncio